Prevent infinite loop crash
If a plugin opens an inventory to a player while they are viewing a smartinv that is non-closeable smartinv will reopen its inventory. This causes the other plugin's inventory to close which causes another close inventory event in smartinv to reopen its inventory. When it reopens another close inventory event will occur even though the smartinv inventory is already opened for the player.