try/finally block for the even passing logic
Since the execution is handed to the end-plugin, it may cause unexpected exceptions. In the previous implementation, if the listener fails, then the plugin is no longer usable since following will happen:

If inventory is closed, close event listener is registered, and the listener fails, then if the inventory is set to closable, the update loop will be keep running despite player's inventory is already closed. This will spam the console and consumes lots of the server resources.

Eventually, it's the end-plugins' job to fix the bug if the listener fails, yet we do not want SmartInv to fail entirely just because of an end-plugin throwing an exception.